requirements.txt
joblib==1.4.0
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
wfdb==4.1.2
Pillow==10.3.0

# image-kit
beautifulsoup4==4.12.2
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.4
requests==2.21.0
scikit-image==0.20.0
seaborn==0.12.2
spacy==3.0.8
validators==0.18.2
html5lib==1.1